Delete de-authenticated users from the db
Also store the GW2 accont the API-key is generated
for alongside. Currently, no unique check is applied.
That opens the possibility to users to hand out
API-keys to foreign users.
But it also allows users who have forgotten their login
and join discord with a new account to re-authenticate
themselves.
The additional field can be used for occasional
aggregate queries over the database to find
potential offenders.