Switched out gw2api, wrapped winston logging into util, started work on multi-server-api-keys
cthos-gw2api could not read logs. The new api can apparently fetch
every endpoint.
Winston always bugged me with its warnings about not having a proper
sink.
It now logs to console with timestamp and proper formatting and to a
file.
With the upcoming link for Riverside, we need a way to allow users
from other server (temporarily) on our discord. In the future,
the config will hold tuples of worlds and discord roles to assign.