Add patch system
Patches are now written as module in the patches/ directory
following a strict interface. This allows for easier application
and rollback of changes in the database schema.
The two existing patches have already been converted into modules.
The application now accepts command line parameters to control
whether the app is executed or patches are applied.