Finish tests for date function and restructure
During testing it became apparent that the project
was probably not properly configured in some parts.
It may also have had to do with a new TSC version, but
the old config did not work out anymore.
In consequence, the entire source project must be contained
within one root directory (src/). That breaks the relative
paths to the config.json, which resided in the root directory
up to this point.
The config, which is taken from the template which is still in
the root directory, must now be put in the src/ directory.
Since this requires the codebase to be recompiled when the config
changes (gulp!) it is no longer feasible to check in the built/
directory, which is therefore ignored from now on.
Instructions as to how to set up the project from now on follow
in a follow-up commit.