Fully implement Glimmer
Glimmer is now fully achievable, minus some bug fixes
and security nets.
To do so, TS3Listener is no longer a discord "listener",
but a fully instantiated class created by the BotgartClient
and accessible through it as well. That is important to enable
other objects (Achievements) to listen to it to trigger
their own evaluation.