Log state of Spawn again, to make sure tick is calcuated for all teams
When ignoring spawns, a faction can have not a single objective on a
map. Summing up their tick over map then yields an empty result.
By having the spawn written to the DB (which is worth 0 points),
each faction produces a valid tick of at least 0 for every map
in each snapshot.